refactor: extract transcript scroll from AgentWorkspace
